fix(android): Convert Float to Double for JSONObject.put() calls
Android's JSONObject only has put(String, double), not put(String, float).
Kotlin's Float maps to primitive float at runtime, causing NoSuchMethodError.

Fixed by explicitly converting Float to Double in createShotMessage().

+++ NativePlugins/Android/GC2AndroidPlugin/src/main/kotlin/com/openrange/gc2/GSProClient.kt
@@ -211,14 +211,15 @@ class GSProClient {
         sideSpin: Float,
         spinAxis: Float
     ): String {
+        // Note: JSONObject.put() requires Double, not Float
         val ballData = JSONObject().apply {
-            put("Speed", ballSpeed)
-            put("SpinAxis", spinAxis)
-            put("TotalSpin", totalSpin)
-            put("BackSpin", backSpin)
-            put("SideSpin", sideSpin)
-            put("HLA", direction)
-            put("VLA", launchAngle)
+            put("Speed", ballSpeed.toDouble())
+            put("SpinAxis", spinAxis.toDouble())
+            put("TotalSpin", totalSpin.toDouble())
+            put("BackSpin", backSpin.toDouble())
+            put("SideSpin", sideSpin.toDouble())
+            put("HLA", direction.toDouble())
+            put("VLA", launchAngle.toDouble())
         }
 
         val options = JSONObject().apply {
